ISDB-320: Transformation implementation
