+++ src/main/resources/application-a58-transform.yaml
@@ -75,6 +75,33 @@ uk:
               courtpreferredDates: court_preferred_dates
               courtInterpreterAssistanceRequired: court_interpreter_assistance_required
               specialAssistanceFacilitiesRequired: special_assistance_facilities_required
+              children:
+                gender: child_gender
+                dateOFBirth: child_dateOFBirth
+                approximateDateOfBirth: child_approximateDateOfBirth
+                placeAndCountry: child_placeAndCoutryOfBirth
+                nationality: child_nationality
+                additionalNationalities: child_nationality
+                isMarried: child_isMarried
+              adopAgencyOrLAs: # if null check it required then implement it in java class
+                adopAgencyOrLaName: child_laName
+                adopAgencyOrLaaddress: child_laAddress
+                adopAgencyOrLaContactName: child_laContactName
+                adopAgencyOrLaPhoneNumber: child_laTelephoneNo
+                adopAgencyOrLaContactEmail: child_laEmail
+              maintanenceOrder:
+                name: child_personName
+                address: child_personAddress
+                nameOfCourt: child_courtAndDateOfOrder.
+                dateOfOrder: child_courtAndDateOfOrder
+                agreementDate: child_agreementDate
+              dateChildMovedIn: child_liveAtHomeFrom
+              hasAnotherAdopAgencyOrLA: child_isAdoptedByAgency
+              proceedingDetails:
+                typeOfOrder: child_typeOfOrder
+                dateOfOrder: child_dateOfOrder
+                nameOfCourt: child_nameOfCourt
+                caseNumber: child_caseNumber
           A58_STEP_PARENT:
             caseFields:
               CASE_TYPE_ID: A58
